Dictionary Added
The dictionary is now added to a website. This button has some of the features.
This has the Search where a user can search for a word and if the word is there it will highlight and bounce to it and if its not an error will appear. There is an edit word feature that will appear a pop up box will appear at the bottom. From then a user will see the word and definition and be able to edit it. There also is an add word feature and a delete word feature.